Supports non-UTC for reading; Do not supports non-UTC for writing
