+++ integration_tests/src/main/python/orc_test.py
@@ -1108,11 +1108,6 @@ def test_orc_non_utc_timezone(reader_confs, end_timestamp, spark_tmp_path, v1_en
     with_cpu_session(lambda spark: gen_df(spark, date_timestamp_gens).write.orc(cpu_write_path), conf=write_confs)
     assert_gpu_and_cpu_are_equal_collect(read_orc_df(cpu_write_path), conf=read_confs)
 
-    # write on GPU
-    gpu_write_path = spark_tmp_path + "/ORC_DATA_GPU"
-    with_gpu_session(lambda spark: gen_df(spark, date_timestamp_gens).write.orc(gpu_write_path), conf=write_confs)
-    assert_gpu_and_cpu_are_equal_collect(read_orc_df(gpu_write_path), conf=read_confs)
-
 @pytest.mark.skip(reason='https://github.com/NVIDIA/spark-rapids/issues/13272: CPU can not read ORC file generated by GPU when timestamp is less than 1970 year')
 @allow_non_gpu(*unsupported_tz_allow_fallback_ops)
 def test_orc_write_but_cpu_read_fail(spark_tmp_path):
@@ -1130,3 +1125,12 @@ def test_orc_write_but_cpu_read_fail(spark_tmp_path):
     with_gpu_session(lambda spark: gen_df(spark, [("c1", ts_gen)]).repartition(1).write.orc(gpu_write_path))
     # Read timestamp on CPU and GPU
     assert_gpu_and_cpu_are_equal_collect(read_orc_df(gpu_write_path))
+
+# Test write with New York timezone and read with different timezones via `tz_sensitive_test` annotation.
+@tz_sensitive_test
+def test_orc_read_write_with_different_timezones(std_input_path):
+    # `timestamp_written_with_New_York_timezone.orc` file is written with New York timezone.
+    # The `writerTimezone` in `StripeFooter` is "America/New_York".
+    # Refer to section `Stripe Footer` in https://orc.apache.org/specification/ORCv1/
+    data_path = f"{std_input_path}/timestamp_written_with_New_York_timezone.orc"
+    assert_gpu_and_cpu_are_equal_collect(lambda spark: spark.read.orc(data_path))
